HVespoli - Cancelar reserva, cambios en algunas variables, registrar Reserva Asignatura

+++ GestorReservasAulas/src/gestionReservasAulas/dominio/CursoExtension.java
@@ -2,9 +2,9 @@ package gestionReservasAulas.dominio;
 
 public class CursoExtension extends Curso {
     private int cantidadClases;
-    private double costoPorAlumno;
+    private float costoPorAlumno;
 
-    public CursoExtension(String nombre, String descripcion, int cantidadAlumnos, int cantidadClases, double costoPorAlumno) {
+    public CursoExtension(String nombre, String descripcion, int cantidadAlumnos, int cantidadClases, float costoPorAlumno) {
         super(nombre,descripcion,cantidadAlumnos);
         this.cantidadClases = cantidadClases;
         this.costoPorAlumno = costoPorAlumno;
